Implementation du modele etapes
