+++ pubspec.yaml
@@ -39,9 +39,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  # Ajout du package pour la gestion des données
+  #Ajout du package pour la gestion des données
   sqflite: ^2.0.0
-
+  path: ^1.8.0
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
